+++ customer-management/src/main/java/com/netcracker/edu/kulich/dao/DefaultCustomerDAO.java
@@ -2,13 +2,12 @@ package com.netcracker.edu.kulich.dao;
 
 import com.netcracker.edu.kulich.entity.Customer;
 import org.springframework.stereotype.Repository;
-import org.springframework.transaction.annotation.Transactional;
 
 import javax.persistence.EntityManager;
+import javax.persistence.EntityNotFoundException;
 import javax.persistence.PersistenceContext;
 import java.util.List;
 
-@Transactional
 @SuppressWarnings("ALL")
 @Repository(value = "customerDAO")
 public class DefaultCustomerDAO implements CustomerDAO {
@@ -41,10 +40,11 @@ public class DefaultCustomerDAO implements CustomerDAO {
         return customer;
     }
 
-
     @Override
-    public void deleteById(Long id) {
-        Customer customer = entityManager.getReference(Customer.class, id);
+    public void deleteById(Long id) throws EntityNotFoundException {
+        Customer customer = entityManager.find(Customer.class, id);
+        if (customer == null)
+            throw new EntityNotFoundException();
         entityManager.remove(customer);
     }
 }
