+++ inventory/src/main/java/com/netcracker/edu/kulich/dao/DefaultOrderItemDAO.java
@@ -2,13 +2,12 @@ package com.netcracker.edu.kulich.dao;
 
 import com.netcracker.edu.kulich.entity.OrderItem;
 import org.springframework.stereotype.Repository;
-import org.springframework.transaction.annotation.Transactional;
 
 import javax.persistence.EntityManager;
+import javax.persistence.EntityNotFoundException;
 import javax.persistence.PersistenceContext;
 import java.util.List;
 
-@Transactional
 @Repository(value = "orderitemDAO")
 public class DefaultOrderItemDAO implements OrderItemDAO {
 
@@ -43,8 +42,11 @@ public class DefaultOrderItemDAO implements OrderItemDAO {
     }
 
     @Override
-    public void delete(Long id) {
-        OrderItem orderItem = entityManager.getReference(OrderItem.class, id);
+    public void delete(Long id) throws EntityNotFoundException {
+        OrderItem orderItem = entityManager.find(OrderItem.class, id);
+        if (orderItem == null) {
+            throw new EntityNotFoundException();
+        }
         entityManager.remove(orderItem);
     }
 }
